refactor(ProductionPage): extract 4 hooks to reduce 865-line megacomponent
Extracts four focused hooks from ProductionPage.tsx (865 → 593 lines, -31%):

- useModalState: consolidates 7 modal open/close booleans
- usePlaybackState: elapsed time, count-in logic, handlePlay/handlePlayWithSpeedUp
- useStickingState: sticking setup toggle + sticking change + applyToSimilar
- useMeasureCopyPaste: measure clipboard with deep-copy clone logic

Also extracts findSimilarMeasures/applyStickingToSimilar pure functions to
src/core/stickingUtils.ts and re-exports them from ProductionPage for backward
compatibility (sticking.test.ts imports from that path).

Closes #126
